requirements.txt
# Google AI & Agent Development Kit
google-adk>=0.3.0
google-generativeai>=0.8.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Google API clients
google-api-python-client>=2.120.0

# AlloyDB / PostgreSQL
asyncpg>=0.29.0
pg8000>=1.31.0
google-cloud-alloydb-connector[asyncpg]>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# MCP (Model Context Protocol)
mcp>=1.3.0

# Async support
anyio>=4.3.0
httpx>=0.27.0

# Configuration & utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Logging & observability
structlog>=24.1.0

# CLI
click>=8.1.0
rich>=13.7.0
